Fixed Point Solvers

Methods and Interface for building solvers for the quadratic team decision problem using empirical fixed point iterations.

Functions

QuadraticTeamDecisionSolver.generateCrossSamples — Method

generateCrossSamples(N::Int, i::Int, γ::Vector{<:Function}, Y::Vector{<:Vector}, R::Vector{<:Vector})

Generate cross samples for a quadratic team decision problem.

Arguments

N::Int: Total number of agents in the problem.

i::Int: Index of the current agent.

γ::Vector{<:Function}: Vector of function class instances representing the decision variables.

R::Vector{<:Vector}: Vector of vectors representing the data samples.

Returns

crossSamples::Vector: Vector of cross samples for the current agent.

QuadraticTeamDecisionSolver.parallelIterationSolver — Method

parallelIterationSolver(
	p::QuadTeamProblem,
	S::Vector{<:Sample},
	functionClass::Vector{<:Function},
	functionNorms::Vector{<:Function},
	outputMap::Function,
	interpolation::Function;
	iterations = 5,
	preprocessData = true,
)

Approximate the solution to a quadratic team decision problem using a parallel iteration scheme.

Arguments

p::QuadTeamProblem: A QuadTeamProblem object defining the problem's specifications.

S::Vector{<:Sample}: A vector of Sample objects representing the data samples.

functionClass::{<:Function}: A vector of functions that for each agent provides a function that constructs an instance of the function class used for optimization.

functionNorm::{<:Function}: A vector of functions that for each agent provides a function that computes the norm of a function class instance.

outputMap::Function: A function that maps performs conditional mean estimation.

interpolation::Function: A function that performs interpolation for updating the decision variables.

iterations::Int = 5: The number of fixed-point iterations to perform.

preprocessData::Bool = true: Whether to perform data preprocessing and validation.

Returns

weights::Vector: A vector of weight vectors representing the decision variables at each iteration.

norms::Vector{Vector{Float64}}: A vector of vectors containing the norms of the weight vectors at each iteration for each agent.

empiricalCost::Vector{Float64}: A vector containing the empirical cost at each iteration.

QuadraticTeamDecisionSolver.GeneralOutputMap — Method

GeneralOutputMap(
	Y::Vector{<:Vector},
	crossSamples::Vector,
	squareBlocks::Vector,
	r::Vector,
	conditionalMean::Function
)

Computes the output map for the agent $i$ in the current iteration.

Arguments:

Samples of $\mathbf{Y}_i$

Samples of $\mathbf{R}_{i,j}(\mathbf{X})\gamma_{(k)}^j(\mathbf{Y}_j)$

Samples of $\mathbf{R}_{i,j}$

Samples of $\mathbf{r}_{i}$

conditionalMean:A method to compute a contional mean of the form $\mathbb{E}[\mathbf{X}|\mathbf{Y}=y]$

Usage:

Pass an anonymous function of the form:

(Yᵢ, Rᵢⱼ, Rᵢᵢ, rᵢ) -> GeneralOutputMap(Y, Rᵢⱼ, Rᵢᵢ, rᵢ, conditionalMean)

as the outputMap argument to any of the fix point solvers.

Note:

Check / fine tune argument types!

Detailed Description of Function Arguments The Solver Takes

functionClass::Function

Arguments

weights: A vector that correspond to the basis function weights / function class parameters

Returns

$f$: A function in the hypothesis function class $\mathcal{F}$.

Remarks

This function is used to construct each entry of the solution function vector γ, as such, great care must be taken that it can properly handle vector valued outputs. Moreover the interpolation and functionNorm methods must be tailor made to suit this function class.

Note that during each iteration, samples of the random variable $\mathbf{R}_{i,j}\gamma_{(k)}^j(\mathbf{Y}_j)$ are approximately generated using the last estimate of $\gamma^j$, so a very good choice of $\mathcal{F}$ is critical to the success of the scheme.

functionNorm::Function

Arguments

weights: A vector that correspond to the basis function weights / function class parameters.

Returns

$||f||$: The function norm of the function $f$ described by the weights weights.

interpolation::Function

Arguments

Y[i]: Measurement data for the agent $i$.

O[i]: Output map data for the agent $i$.

Returns

weights: The weights that correspond to the function $f$ in our function class $\mathcal{F}$ that best interpolates Y[i] and O[i].

Remarks

As with functionClass, care most be taken that this can hanble both vector and scalar valued input and output data.

outputMap::Function

Arguments

Y[i]: Measurement data for the agent $i$.

crossSamples: generated sample data from the random variable given by: $\mathbf{R}_{i,j}\gamma_{(k)}^j(\mathbf{Y}_j)$.

squareBlocks: Samples from $\mathbf{R}_{i,i}$. These should be Hermitian and Positive Definite.

r_blocks[i]: Samples from $\mathbf{r}_i$

Returns

O: Vector of approximated samples of the random variable $\left( \mathbb{E}[\mathbf{R}_{i,i}|\mathbf{Y}_i] \right)^{-1}\left( \sum_{i,j=1 \ , \ i \neq j}^N \mathbb{E}[\mathbf{R}_{i,j}\gamma_{(k)}^j(\mathbf{Y}_j)|\mathbf{Y}_i] + \mathbb{E}[\mathbf{r}_i|\mathbf{Y}_i] \right)$

Remarks

This method must have within it a method that estimates conditional expectations of the form $\mathbb{E}[\mathbf{X}|\mathbf{Y}=y]$ see: GeneralOutputMap(Y::Vector{<:Vector}, crossSamples::Vector{<:Vector}, squareBlocks::Vector{<:Matrix}, r::Vector{<:Vector}, conditionalMean::Function)

Otherwise, an outputMap can also be implemented with a method that estimates conditional expectation operators of the form $\mathbb{E}[\mathbf{X}|\mathbf{Y}]$.