Scalar and Vector Kernel Methods

The solvers implemented in this library reproducing kernel Hilbert space methods to approximate the conditional expectation functions that arise in our iterations.

For this reason, we provide a flexible and easy to use interface that can be employed to solve both scalar and vector valued regression problems.

We only provide two kernel definitions, one for the scalar case and one for the vector case, respectively. Nevertheless, the user can easily define their own kernels and use them. We make no requirement on the type and behaviour of a kernel, other that it must be a function.

For more details on the general theory of scalar reproducing kernel Hilbert space methods, please refer to (Berlinet and Thomas-Agnan, 2004).

For more details on the formulation of vector reproducing kernel Hilbert space methods, please refer to (Micchelli and Pontil, 2005), for details on approximation properties and universality of these spaces refer to (Carmeli et al., 2008).

Functions

QuadraticTeamDecisionSolver.exponentialKernel — Method

exponentialKernel(x::Vector, y::Vector; h = 1)

Radial Basis Function Kernel

Arguments:

x::Vector: A data sample.

y::Vector: Also a data sample.

h=1: Width of the kernel.

Returns:

Value of the kernel computed at x and y with window width h,

that is, $\exp \left( - \frac{|x - y|^2}{h} \right)$ where $| \ \cdot \ |$ is the norm on the original sample space.

QuadraticTeamDecisionSolver.matrixExponentialKernel — Method

matrixExponentialKernel(h::Vector{Float64}, M::Vector{<:Matrix}, λ::Vector{Float64}, x::Any, t::Any)

Matrix Valued Gaussian Mixture Kernel

Arguments:

h::Vector{Float64}: Vector of kernel bandwidths $\sigma_i, i \in \{ 1, \dots, L \}$.

M::Vector{<:Matrix}: Vector of mixture matrices $M_i, i \in \{ 1, \dots, L \}$.

λ::Vector{Float64}: Vector of convex combination coefficients $\lambda_i,i \in \{ 1, \dots, L \}^2, \ \sum_{i=1}^L\lambda_i = 1$.

x::Vector: A data sample $x \in \mathcal{X}$.

x::Vector: Also a data sample.

Returns:

Value of matrix kernel computed at x and y with specified input parameters.

\[ K(x,t) = \sum_{i=1}^L \lambda_ie^{-\sigma_i^{-1}d_{\mathcal{X}}(x,y)^2}M_i\]

QuadraticTeamDecisionSolver.gramian — Method

gramian(kernel::Function, X::Vector{<:Vector})

Compute the gramian of a kernel $K$ over the samples $\mathbf{X}^n \subset \mathcal{X}$.

Arguments:

kernel::Function: A scalar or matrix kernel function.

X::Vector{<:Vector}: The samples over which the gramian is to be constructed.

QuadraticTeamDecisionSolver.kernelNorm — Method

kernelNorm(weights::Vector, kernelGramian::Matrix)

Compute the function norm: $\sum_{i=1}^n \sum_{j=1}^n \mu_i \mu_j K( x^{(i)}, x^{(j)} )$ of a function in a reproducing kernel Hilbert space. Th

Arguments:

weights::Vector: The $\mu$'s in the above expression.

kernelGramian::Matrix: The gramian of $K$ over the samples $\mathbf{X}^n$

QuadraticTeamDecisionSolver.kernelFunction — Method

kernelFunction(
	kernel::Function,
	weights::Vector,
	X::Vector{<:Vector},
	x::Vector,
)

Evaluates a kernel function of the form $f(x) = \sum_{i=1}^n \mu_i K(x, x^{(i)})$

Arguments:

kernel::Function scalar or matrix valued Kernel function that defines the RKHS where $f$ lives.

weights: The $\mu$'s in the above expression.

x: The point $\mathbf{x} \in \mathcal{X}$ at which $f$ is to be evaluated.

QuadraticTeamDecisionSolver.kernelRegression — Method

kernelRegression(
	kernel::Function,
	Y::Vector,
	X::Vector;
	λ = 0.5,
)

Solves the kernel regression problem

\[f^* = \argmin_{f \in \mathcal{H}_k} \sum_{i=1}^n \lvert \lvert f(x^{(i)}) - y^{(i)} \rvert \rvert_{\mathcal{Y}}^2 + \lambda \lvert \lvert f \rvert \rvert_{\mathcal{H}_k}^2 \]

Where $\{x^{(i)}, y^{(i)} \}_{i=1}^n \subset \mathcal{X}\times\mathcal{Y}$ are data samples, and $\mathcal{H}_k$ is a Reproducing Kernel Hilbert Space with kernel $k$.

Moreover, this implementation can handle the case $\mathcal{Y} = \mathbb{C}^d$ for $d > 1$. That is, this implements vector valued kernel regression, as well as scalar valued kernel regression.

Arguments:

kernel::Function: The kernel function that corresponds to $\mathcal{H}_k$.

X::Vector: Vector of samples in input space $\mathcal{X}$.

Y::Vector: Vector of samples in output space $\mathcal{Y}$.

λ = 0.5: Regularization constant.

Note that Y::Vector and X::Vector must have the same length n.