General Types and Utility Functions

Types

QuadraticTeamDecisionSolver.QuadTeamProblem — Type

QuadTeamProblem{T <: Number}

The QuadTeamProblem struct represents a quadratic team decision problem. In essence, it stores the correct problem dimensions and the field over which the problem is solved.

Fields

N::Int: Number of agents in the team.

m::Vector{Int}: Array of measurement dimensions for each agent.

a::Vector{Int}: Array of action space dimensions for each agent.

T::Type{T}: Numeric type for the problem.

QuadraticTeamDecisionSolver.Sample — Type

Sample{T <: Number}

The Sample struct represents a sample data point for the quadratic team decision problem.

Fields

Y::Vector{Vector{T}}: Array of measurement vectors for each agent, where Y[i] represents the measurement vector for agent i.

R::Matrix{T}: Quadratic matrix capturing the interaction between agents.

r::Vector{T}: Vector representing the linear term for the control policies.

c::T: Constant term for the objective function.

Sample Data

In the quadratic team decision problem, a sample represents a specific data point with measurements, quadratic matrices, and linear and constant terms. These samples are used to estimate the performance and optimize the control policies.

The sample data includes the following components:

$\mathbf{Y}$: Array of measurement vectors for each agent, where Y[i] represents the measurement vector for agent i.

$\mathbf{R}$: Hermitian Positive Definite matrix capturing the interaction between agents. It represents the quadratic term for the control policies.

$\mathbf{r}$: Vector representing the linear term for the control policies.

$\mathbf{c}$: Constant term for the objective function.

Functions

QuadraticTeamDecisionSolver.checkProblem — Method

checkProblem(p::QuadTeamProblem)

Check the consistency and correctness of a QuadTeamProblem object p.

Arguments

p::QuadTeamProblem: The QuadTeamProblem object representing the problem specification.

Errors

AssertionError: Throws an error if the problem specification is incorrect or inconsistent.

Example

P = QuadTeamProblem(...)
checkProblem(P)

QuadraticTeamDecisionSolver.checkSample — Method

checkSample(p::QuadTeamProblem, s::Sample)

Check the consistency and correctness of a Sample object s with respect to a QuadTeamProblem p.

Additionally checks samples of $\mathbf{R}(\mathbf{X})$ are Hermitian and positive definite.

Arguments

p::QuadTeamProblem: The QuadTeamProblem object representing the problem specification.

s::Sample: The Sample object to be checked.

Errors

AssertionError: Throws an error if any inconsistency or mismatch is found in the sample, or if R isn't positive definite

Example

P = QuadTeamProblem(...)
s = Sample(...)
checkSample(P, s)

QuadraticTeamDecisionSolver.checkGamma — Method

checkGamma(P::QuadTeamProblem, γ::Vector{<:Function})

Check the output dimensions of the functions γ for each agent in the QuadTeamProblem P.

Arguments

P::QuadTeamProblem: The QuadTeamProblem specifying the problem.

γ::Vector{Function}: Vector of functions γ for each agent.

Returns

γ::Vector{Function}: The input vector of functions γ.

Errors

Throws an error if the output dimensions of the functions γ do not match the specified dimensions in P.

QuadraticTeamDecisionSolver.checkData — Method

checkData(p::QuadTeamProblem, S::Vector{<:Sample})

Check the validity of a vector of Sample data for a given QuadTeamProblem p.

Arguments

p::QuadTeamProblem: The QuadTeamProblem specifying the problem.

S::Vector{Sample}: Vector of Sample data to be checked.

Returns

result::Vector{Sample}: A vector of Sample data with validity checks applied.

QuadraticTeamDecisionSolver.loss — Method

loss(s::Sample, γ::Vector{<:Function})

Compute the loss function for a given Sample s using a vector of functions γ.

Arguments

s::Sample: The Sample data.

γ::Vector{Function}: Vector of functions specifying the control policies.

Returns

loss::Real: The computed loss value.

QuadraticTeamDecisionSolver.risk — Method

risk(S::Vector{<:Sample}, γ::Vector{<:Function})

Compute the risk function for a given vector of Samples S using a vector of functions γ.

Arguments

S::Vector{Sample}: Vector of Sample data.

γ::Vector{Function}: Vector of functions specifying the control policies.

Returns

risk::Real: The computed risk value.

QuadraticTeamDecisionSolver.splitSampleIntoBlocks — Method

splitSampleIntoBlocks(p::QuadTeamProblem, s::Sample)

Split a sample s into blocks based on the dimensions defined in the QuadTeamProblem p.

Arguments

p::QuadTeamProblem: A QuadTeamProblem object defining the dimensions of the blocks.

s::Sample: A Sample object to be split into blocks.

Returns

R_blocks::Vector{SubArray}: An array of subarrays containing the blocks of s.R based on the block dimensions defined in p.

r_blocks::Vector{SubArray}: An array of subarrays containing the blocks of s.r based on the block dimensions defined in p.

QuadraticTeamDecisionSolver.splitDataSetIntoBlocks — Method

splitDataSetIntoBlocks(p::QuadTeamProblem, S::Vector{<:Sample})

Split a vector of samples S into blocks based on the dimensions defined in the QuadTeamProblem p.

Arguments

p::QuadTeamProblem: A QuadTeamProblem object defining the dimensions of the blocks.

S::Vector{<:Sample}: A vector of Sample objects to be split into blocks.

Returns

Y: vector of length N (number of agents) of vectors Y[i] of length m (number of samples).

Each entry of Y is the vector of measurement vectors that correspond to the agent i.

R_blocks: vector of length N (number of agents) of vectors R of length m (number of samples).

Each entry of R, R[i] is a vector of the blocks of R that correspond to the agent i

r_blocks: vector of length N (number of agents) of vectors r of length m (nummber of samples).

Each entry of r, r[i] is a vector of the block of r that corresponds to the agent i