General Types and Utility Functions

These types and functions provide basic tools for defining and dealing with static quadratic team problems.

Types

QuadraticTeamDecisionSolver.QuadTeamProblem — Type

QuadTeamProblem{T <: Number}

The QuadTeamProblem struct represents a quadratic team decision problem. In essence, it stores the correct problem dimensions and the field over which the problem is solved.

Fields

N::Int: Number of agents in the team.

m::Vector{Int}: Array of measurement dimensions for each agent.

a::Vector{Int}: Array of action space dimensions for each agent.

T::Type{T}: Numeric type for the problem.

Functions

QuadraticTeamDecisionSolver.checkProblem — Method

checkProblem(p::QuadTeamProblem)

Check the consistency and correctness of a QuadTeamProblem object p.

Arguments

p::QuadTeamProblem: The QuadTeamProblem object representing the problem specification.

Errors

AssertionError: Throws an error if the problem specification is incorrect or inconsistent.

Example

P = QuadTeamProblem(...)
checkProblem(P)

QuadraticTeamDecisionSolver.checkGamma — Method

checkGamma(p::QuadTeamProblem, γ::Vector{<:Function})

Check the output dimensions of the functions γ for each agent in the QuadTeamProblem P.

Arguments

P::QuadTeamProblem: The QuadTeamProblem specifying the problem.

γ::Vector{Function}: Vector of functions γ for each agent.

Returns

γ::Vector{Function}: The input vector of functions γ.

Errors

Throws an error if the output dimensions of the functions γ do not match the specified dimensions in P.

QuadraticTeamDecisionSolver.residual — Method

residual(
	m::Int,
	p::QuadTeamProblem,
	kernels::Vector{<:Function},
	γ::Vector{<:Vector{<:Vector}},
	Y::Vector{<:Vector},
	Q::Matrix{<:Vector},
	R::Vector{<:Vector},
	λ::Vector{<:AbstractFloat},
)

Compute the residual function

\[\mathrm{res}(\gamma) := \mathbf{A}\gamma + \mathbf{\tilde{R}}\]

for a given tuple of policies $\gamma = (\gamma^1, \dots, \gamma^N)$.

Arguments

m::Int: Number of training samples.

p::QuadTeamProblem: A quad team problem object.

kernels::Vector{<:Function}: Vector of kernel functions.

γ::Vector{<:Vector{<:Vector}}: Vector of gamma kernel function coefficients.

Y::Vector{<:Vector}: Vector of measurement vector samples.

Q::Matrix{<:Vector}: Matrix of system matrix samples, organized block-wise.

R::Vector{<:Vector}: Vector of linear term samples, organized block-wise.

Returns

res: vector of kernel function coefficient vectors corresponding to residual function.

QuadraticTeamDecisionSolver.gammaNorm — Method

gammaNorm(f::Function, Y::AbstractVector)

Compute the $\Gamma^i$ norm of a given function $f \in \Gamma^i$ approximated using measurement vector data, that is, samples of $\mathbf{Y}_i$.

Arguments

f::Function: The function for which the norm is to be computed.

Y::AbstractVector: Input data vector.

Returns

norm: $\Gamma^i$-norm of f.

Details

The gamma norm is calculated as the squared Euclidean norm of the function's evaluations on the input data, normalized by the length of the input data.

QuadraticTeamDecisionSolver.GammaNorm — Method

GammaNorm(F::Vector{<:Function}, Y::AbstractVector)

Compute the $\Gamma$ norm of a given function tuple $F = (f_1, \dots, f_i) \in \Gamma = \Gamma^i\times\dots\times\Gamma^N$ approximated using measurement vector data, that is, samples of $\mathbf{Y} = (\mathbf{Y}_i,\dots,\mathbf{Y}_N)$.

Arguments

F::Vector{<:Function}: Vector of functions for which the gamma norm is to be computed.

Y::AbstractVector: Input data vector.

Returns

norm: $\Gamma$-norm of F.