Team Mininum Mean Square Error Precoding

Here we provide a simple interface and sample generation tool for producing a toy simulation that we can test our solvers on.

General Team Minimum Mean Square Error Problem Description

The team MMSE precoding problem is in general a static quadratic team decision problem. In particular, if we have $N$ transmitters each equipped with $L$ antennas and $K$ single antenna receivers, then for every $k \in \{1,\dots, K\}$, the team MMSE problem reads

\[\argmin_{\gamma_k \in \Gamma^1 \times \dots \times \Gamma^N} \mathrm{MSE}_k(\gamma) := \mathbb{E} \left[ \left\lVert \mathbf{H}^{\mathsf{H}}\gamma_k(\mathbf{Y}) - \mathbf{e}_k \right\rVert^2 + \frac{1}{P}\left\lVert \gamma_k(\mathbf{Y}) \right\rVert^2 \right].\]

The random variable

\[\mathbf{H} = [\mathbf{H}_1, \dots, \mathbf{H}_N]\]

is called the channel and is centrally symmetric complex gaussian with realizations in $\mathbb{C}^{K, LN}$ and arbitrary covariance.

For each transmitter (agent) $i \in \{1,\dots, N\}$ , the measurement random vector $\mathbf{Y}_i$ is given by

\[\mathbf{Y}_i := \mathbf{H} + \mathbf{Z}_i\]

where $\mathbf{Z}_i$ is the channel estimation error at transmitter (agent) $i$ and is also centrally symmetric complex gaussian with realizations in ``\mathbb{C}^{K,LN} and arbitrary covariance.

For further details on team MMSE precoding, please refer to (Miretti et al., oct 2021) and (Miretti et al., 2022).

Our Situation

The previously defined measurement model is quite general. While the solvers we provide should be able to handle all measurement schemes that can be constructed from this model in this library, we implement the following model:

\[\mathbf{Y}_i = \begin{bmatrix} \dots & \varepsilon\mathbf{H}_{i-1} + (1 - \varepsilon)\mathbf{N}_{i-1} & \mathbf{H}_i & \varepsilon\mathbf{H}_{i+1} + (1 - \varepsilon)\mathbf{N}_{i+1} & \dots \end{bmatrix}\]

where $\mathbf{H}_i$ is the $i$-th block column of the channel $\mathbf{H}$ and $\mathbf{N}$ is centrally symmetric gaussian noise with diagonal covariance matrix.

Types

QuadraticTeamDecisionSolver.teamMMSEproblem — Type

struct teamMMSEproblem

A struct to specify parameters for a problem from the team MMSE class.

Attributes:

P::AbstractFloat: The Signal-to-Interference-plus-Noise Ratio (SINR) for the problem.

N::Int: The number of transmitters in the problem.

L::Int: The number of antennas at each transmitter.

K::Int: The number of receivers in the problem.

σₕ::Vector{AbstractFloat}: Vector specifying channel covariances.

σₙ::Vector{AbstractFloat}: Vector specifying noise covariances.

ϵ::Vector{AbstractFloat}: vector of link coefficients $\varepsilon_i \in (0,1), i \in \{1,\dots,N\}$.

Functions

QuadraticTeamDecisionSolver.sampleComplexNormal — Method

sampleComplexNormal(σ::AbstractFloat, K::Int, L::Int)

Sample from the circularly symmetric complex Gaussian distribution.

Parameters:

σ::AbstractFloat: The standard deviation of the normal distributions for both the real and imaginary parts of the complex numbers.

K::Int: The number of rows in the output matrix.

L::Int: The number of antennas

Returns:

A KxN matrix of complex numbers with circularly symmetric Gaussian distribution.

QuadraticTeamDecisionSolver.generateTeamMMSEsamples — Method

generateTeamMMSEsamples(t::teamMMSEproblem, m::Int)

generate m samples for a team MMSE problem specified by t.

Arguments

t::teamMMSEproblem: a teamMMSEproblem object specifying the statistics and sizes of channel and noise measurements

m::Int: the desired amount of samples

Returns

H:: A vector of size t.N of vectors of size m of $K \times L$ channel measurement matrices

Y:: A vector of size t.N of vectors of size m of $K \times L \times N$ agent measurement vectors

Q:: A matrix of size t.N by t.N of vectors of size m of $L \times L$ matrices

R:: A vector of size t.N of vectors of size m of $L \times 1$ matrices