Fixed Point Solvers

Methods and Interface for building solvers for the quadratic team decision problem using empirical methods.

Explanation

The solution to the static quadratic team decision problem is entirely characterized by a set of linear operator equations over the (product) policy space. These are called the Stationarity Equations, and for all $i \in \{1, \dots, N\}$ they read

\[\mathbb{E} \left[ \mathbf{Q}_{ii} \vert \mathbf{Y}_i \right]\gamma^i(\mathbf{Y}_i) + \sum_{i,j=1 , j \neq i} \mathbb{E} \left[ \mathbf{Q}_{ij} \gamma_{(k)}^j \left( \mathbf{Y}_j \right) \vert \mathbf{Y}_i \right] + \mathbb{E} \left[ \mathbf{R}_i \vert \mathbf{Y}_i \right] = 0 \text{ a.s.}.\]

We can write these compactly as

\[\mathbf{A}\gamma = - \tilde{\mathbf{R}}.\]

Write

\[\mathbf{A} = \mathbf{A}_1 + \mathbf{A}_2\]

with easily invertible $\mathbf{A}_1$, and consider the iteration

\[\gamma_{k+1} = -\mathbf{A}_1^{-1}\left( \mathbf{A}_2\gamma_{k} + \tilde{\mathbf{R}} \right)\]

which is a fixed point iteration. This iteration is generated by the operator splitting $\mathbf{A} = \mathbf{A}_1 + \mathbf{A}_2$. For this reason, we dub the methods implemented in this library operator splitting methods, which, given our problem structure, are entirely identical to their finite dimensional matric analogues.

For more details on this class of operator splittings, please see (de Pillis, 1973).

Convergence Behavior

The convergence behavior of the methods implemented in this library depends upon the eigenvalues of the system matrix $\mathbf{Q}$. In particular, the assumption that $\mathbf{Q}$ is bounded away from zero implies that the decomposition

\[\mathbf{Q} = \mathbf{T} + \lambda_{\mathrm{min}}I,\]

for positive semidefinite $\mathbf{T}$ is always valid.

The convergence of our schemes then depends upon the quantitoes$\varrho(\mathbf{T})$ and $\lambda_{\mathrm{min}}$. Note that $\varrho(\mathbf{T})$ exists and is finite as, by assumption, $\mathbf{Q}$ is uniformly bounded above.

Functions

QuadraticTeamDecisionSolver.jacobiSolver — Method

jacobiSolver(
	p::QuadTeamProblem,
	m::Int,
	Y::Vector{<:Vector},
	Q::Matrix{<:Vector},
	R::Vector{<:Vector},
	kernels::Vector{<:Function},
	λ::Vector{Float64};
	iterations = 5,
	random_init = false,`
)

Implements the Jacobi operator splitting method to generate policy updates.

Converges to the unique team optimal $\gamma^*$ whenever

\[\lambda_{\mathrm{min}} > \varrho(\mathbf{T})\]

QuadraticTeamDecisionSolver.gaussSeidelSolver — Method

gaussSeidelSolver(
	p::QuadTeamProblem,
	m::Int,
	Y::Vector{<:Vector},
	Q::Matrix{<:Vector},
	R::Vector{<:Vector},
	kernels::Vector{<:Function},
	λ::Vector{Float64};
	iterations = 5,
	random_init = false,
)

Implements the Gauss-Seidel operator splitting method to generate policy updates.

Guaranteed to converge to the unique team optimal $\gamma^*$.

QuadraticTeamDecisionSolver.SORSolver — Method

SORSolver(
	p::QuadTeamProblem,
	m::Int,
	Y::Vector{<:Vector},
	Q::Matrix{<:Vector},
	R::Vector{<:Vector},
	kernels::Vector{<:Function},
	λ::Vector{Float64};
	iterations = 5,
	random_init = false,
	omega = 1.0,
)

Implements the Successive-Over-Relaxation (SOR) operator splitting method to generate policy updates.

Converges to the unique team optimal $\gamma^*$ for all $\omega \neq 0$ that fulfill

\[ 1 - \frac{\lambda_{\mathrm{min}}}{\varrho(\mathbf{T})} < \omega < 1 + \frac{\lambda_{\mathrm{min}}}{\varrho(\mathbf{T})}\]